requirements.txt
# ISP Middleware Flask Application - Requirements
# Place this in your Flask app root directory

# Core Flask
Flask>=2.3.0
Flask-SQLAlchemy>=3.0.0
Flask-Migrate>=4.0.0
Flask-CORS>=6.0.0
Flask-JWT-Extended>=4.5.0
Flask-Limiter>=3.3.0
Flask-Caching>=2.0.0

# Database
psycopg2-binary>=2.9.0  # PostgreSQL
redis>=4.5.0            # Redis for caching
alembic>=1.11.0         # Database migrations

# API and Networking
requests>=2.31.0
python-decouple>=3.8
marshmallow>=3.19.0     # Serialization
webargs>=8.2.0          # Request parsing

# RADIUS Integration
pyrad>=2.4              # RADIUS client/server
cryptography>=41.0.0    # Encryption

# MikroTik API
librouteros>=3.2.0      # MikroTik RouterOS API
paramiko>=3.2.0         # SSH for MikroTik

# OpenVPN Management
#python-openvpn>=1.0.0   # OpenVPN management
subprocess32>=3.5.4     # Process management

# Network Utilities
netaddr>=0.8.0          # IP address manipulation
scapy>=2.5.0           # Packet manipulation
netifaces>=0.11.0      # Network interface info

# Monitoring and Logging
prometheus-client>=0.17.0  # Metrics
structlog>=23.1.0          # Structured logging
python-json-logger>=2.0.7  # JSON logging

# Authentication and Security
bcrypt>=4.0.0           # Password hashing
pyotp>=2.8.0           # Two-factor authentication
python-jose>=3.3.0     # JWT tokens

# Task Queue (optional for background tasks)
celery>=5.3.0          # Background tasks
flower>=2.0.0          # Celery monitoring

# Development and Testing
pytest>=7.4.0
pytest-flask>=1.2.0
factory-boy>=3.3.0
coverage>=7.2.0

# System Integration
psutil>=5.9.0          # System monitoring
schedule>=1.2.0        # Task scheduling

# Configuration Management
pyyaml>=6.0            # YAML config files
configparser>=5.3.0    # INI config files

# Network Monitoring
ping3>=4.0.0           # Network ping
speedtest-cli>=2.1.3   # Speed testing

# API Documentation
flask-swagger-ui>=4.11.0  # Swagger UI
apispec>=6.3.0           # OpenAPI spec generation

gunicorn